Refactor project into modular structure
