+++ newFoodTracker/Services/UserProfileService.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+class UserProfileService: ObservableObject {
+    static let shared = UserProfileService()
+    @Published var currentProfile = UserProfile()
+    private init() {}
+}
