+++ newFoodTracker/Views/Components/CardView.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+struct CardView<Content: View>: View {
+    let content: Content
+    init(@ViewBuilder content: () -> Content) { self.content = content() }
+    var body: some View {
+        content
+            .padding()
+            .background(
+                RoundedRectangle(cornerRadius: 16)
+                    .fill(Color.cardBackground)
+                    .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
+            )
+            .cornerRadius(16)
+    }
+}
