+++ newFoodTracker/Views/HomeView.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+struct HomeView: View {
+    var body: some View {
+        WeeklyPlanOverviewView()
+            .environmentObject(MealPlanService())
+    }
+}
